Try to use chdir to get around the sun_path length limit.
If a socket's path name would overflow sun_path, try to make it fit by
temporarily changing to the directory the socket is in and using the
path's basename in sun_path.

This should help for the case where the basename of the socket does fit
in sun_path, even if the rest of the path doesn't.