Avoid permission race condition when dtach socket is created
Use umask(077) to avoid a race condition between the creation of the
dtach socket and the chmod system call on the socket.

This race condition can potentially be exploited when the dtach socket
is created in a directory such as /tmp.

(cherry picked from commit b71dfb2e2c02dfe14738781214a5734d63eeb772)